+++ app/src/main/java/com/marcinorlowski/fonty/demo/MainActivity.java
@@ -3,7 +3,7 @@ package com.marcinorlowski.fonty.demo;
 /*
  ******************************************************************************
  *
- * Copyright 2013-2017 Marcin Orlowski
+ * Copyright 2013-201 Marcin Orlowski
  *
  * Licensed under the Apache License 2.0
  *
@@ -30,16 +30,22 @@ import android.widget.Toast;
 
 import com.marcinorlowski.fonty.Fonty;
 
+import butterknife.ButterKnife;
+import butterknife.OnClick;
+import butterknife.Unbinder;
+
 public class MainActivity extends AppCompatActivity {
 
 	ActionBarDrawerToggle mDrawerToggle;
+
+	protected Unbinder mButterKnifeUnbinder;
 
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.activity_main);
 
-		Toolbar toolbar = (Toolbar)findViewById(R.id.toolbar);
+		Toolbar toolbar = findViewById(R.id.toolbar);
 		toolbar.setTitle("Fonty by Marcin Orlowski");
 		toolbar.setSubtitle("Easily change fonts of your app!");
 		setSupportActionBar(toolbar);
@@ -47,69 +53,65 @@ public class MainActivity extends AppCompatActivity {
 		final ActionBar actionBar = getSupportActionBar();
 
 		if (actionBar != null) {
-			DrawerLayout drawerLayout = (DrawerLayout)findViewById(R.id.drawer_layout);
+			DrawerLayout drawerLayout = findViewById(R.id.drawer_layout);
 
 			actionBar.setDisplayHomeAsUpEnabled(true);
 			mDrawerToggle = new ActionBarDrawerToggle(this, drawerLayout, toolbar, R.string.toggle_opend, R.string.toggle_closed) {
-				public void onDrawerClosed(View view) {
+				public void onDrawerClosed(View drawerView) {
+					super.onDrawerClosed(drawerView);
 					supportInvalidateOptionsMenu();
 				}
 
 				public void onDrawerOpened(View drawerView) {
+					super.onDrawerOpened(drawerView);
 					supportInvalidateOptionsMenu();
 				}
 			};
 			mDrawerToggle.setDrawerIndicatorEnabled(true);
 
-			drawerLayout.setDrawerListener(mDrawerToggle);
-
+			drawerLayout.addDrawerListener(mDrawerToggle);
 			mDrawerToggle.syncState();
 		}
-
 
 		// apply our font replacement
 		Fonty.setFonts(this);
 
-		int[] ids = {R.id.github, R.id.button};
-		for (int id : ids) {
-			findViewById(id).setOnClickListener(mOnClickListener);
+		mButterKnifeUnbinder = ButterKnife.bind(this);
+	}
+
+	@Override
+	protected void onDestroy() {
+		mButterKnifeUnbinder.unbind();
+		super.onDestroy();
+	}
+
+	@OnClick (R.id.github)
+	public void clickGithub() {
+		try {
+			startActivity(
+					new Intent(Intent.ACTION_VIEW,
+							Uri.parse("https://github.com/MarcinOrlowski/fonty/")
+					)
+			);
+		} catch (ActivityNotFoundException e) {
+			Toast.makeText(MainActivity.this, "Failed to launch Web Browser", Toast.LENGTH_SHORT).show();
 		}
 	}
 
-
 	protected Handler mHandler = new Handler();
 
-	protected View.OnClickListener mOnClickListener = new View.OnClickListener() {
-		@Override
-		public void onClick(View view) {
-			switch (view.getId()) {
-				case R.id.button:
-					int[] ids = {R.id.til1, R.id.til2};
-					for (int id : ids) {
-						TextInputLayout til = (TextInputLayout)findViewById(id);
-						til.setErrorEnabled(true);
-						til.setError("Some error text");
-					}
-
-					mHandler.removeCallbacks(mRunnable);
-					mHandler.postDelayed(mRunnable, 6 * 1000);
-					break;
-
-				case R.id.github:
-					try {
-						startActivity(
-								new Intent(Intent.ACTION_VIEW,
-										Uri.parse("https://github.com/MarcinOrlowski/fonty/")
-								)
-						);
-					} catch (ActivityNotFoundException e) {
-						Toast.makeText(MainActivity.this, "Failed to launch Web Browser", Toast.LENGTH_SHORT).show();
-					}
-					break;
-			}
+	@OnClick (R.id.button)
+	public void clickButton() {
+		int[] ids = {R.id.til1, R.id.til2};
+		for (int id : ids) {
+			TextInputLayout til = findViewById(id);
+			til.setErrorEnabled(true);
+			til.setError("Some error text");
 		}
-	};
 
+		mHandler.removeCallbacks(mRunnable);
+		mHandler.postDelayed(mRunnable, 6 * 1000);
+	}
 
 	// Runnable that hides TIL's error messages
 	protected Runnable mRunnable = new Runnable() {